% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg.CU.R
\name{alg.CU}
\alias{alg.CU}
\title{CU quantile-based clustering algorithm}
\usage{
alg.CU(data, k = 2, eps = 1e-08, it.max = 100, B = 30)
}
\arguments{
\item{data}{A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.}

\item{k}{The number of clusters. The default is k=2.}

\item{eps}{The relative convergence tolerances for objective function. The default is set to 1e-8.}

\item{it.max}{A number that gives integer limits on the number of the CU algorithm iterations. By default, it is set to 100.}

\item{B}{The number of times the initialization step is repeated; the default is 30.}
}
\value{
A list containing the following elements:
\item{method}{The chosen parameterization, CU, Common theta and Unscaled variables}
\item{k}{The number of clusters.}
\item{cl}{A vector whose [i]th entry is classification of observation i in the test data.}
\item{qq}{A matrix whose [h,j]th entry is the theta-quantile of variable j in cluster h.}
\item{theta}{A vector whose [j]th entry is the percentile theta for variable j.}
\item{Vseq}{The values of the objective function V at each step of the algorithm.}
\item{V}{The final value of the objective function V.}
\item{lambda}{A vector containing the scaling factor for each variable.}
}
\description{
This function allows to run the CU (Common theta and Unscaled variables) version of the quantile-based clustering algorithm.
}
\details{
Algorithm CU: Common theta and Unscaled variables. A common value of theta for all the variables is assumed.
This strategy directly generalizes the conventional k-means to other moments of the distribution to better accommodate skewness in the data.
}
\examples{
out <- alg.CU(iris[,-5],k=3)
out$theta
out$qq

table(out$cl)
}
\references{
Hennig, C., Viroli, C., Anderlucci, L. (2019) "Quantile-based clustering" \emph{Electronic Journal of Statistics}, 13 (2) 4849-4883  <doi:10.1214/19-EJS1640>
}
