% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alg.VU.R
\name{alg.VU}
\alias{alg.VU}
\title{VU quantile-based clustering algorithm}
\usage{
alg.VU(data, k = 2, eps = 1e-08, it.max = 100, B = 30)
}
\arguments{
\item{data}{A numeric vector, matrix, or data frame of observations. Categorical variables are not allowed. If a matrix or data frame, rows correspond to observations and columns correspond to variables.}

\item{k}{The number of clusters. The default is k=2.}

\item{eps}{The relative convergence tolerances for objective function. The default is set to 1e-8.}

\item{it.max}{A number that gives integer limits on the number of the VU algorithm iterations. By default, it is set to 100.}

\item{B}{The number of times the initialization step is repeated; the default is 30.}
}
\value{
A list containing the following elements:
\item{cl}{A vector whose [i]th entry is classification of observation i in the test data.}
\item{qq}{A matrix whose [h,j]th entry is the theta-quantile of variable j in cluster h.}
\item{theta}{A vector whose [j]th entry is the percentile theta for variable j.}
\item{Vseq}{The values of the objective function V at each step of the algorithm.}
\item{V}{The final value of the objective function V.}
}
\description{
This function allows to run the VU (Variable-wise theta_j and Unscaled variables) version of the quantile-based clustering algorithm.
}
\details{
Algorithm VU: Variable-wise theta_j and Unscaled variables. A different theta for every single variable is estimated to better accomodate different degree of skeweness in the data.
}
\examples{
out <- alg.VU(iris[,-5],k=3)
out$theta
out$qq

table(out$cl)
}
\references{
C. Hennig, C. Viroli, L. Anderlucci (2018). \emph{Quantile-based clustering}. \url{http://arxiv.org/abs/1806.10403}
}
